pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.18",
    "ccxt>=4.4.30",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "requests>=2.32.3",
]
